+++ src/main/java/cz/cvut/fit/tjv/popovle1/semestral/crud_gamedev/dto/DevDTO.java
@@ -1,13 +1,13 @@
 package cz.cvut.fit.tjv.popovle1.semestral.crud_gamedev.dto;
 
 public class DevDTO {
-    private final Long id;
+    private final Integer id;
     private final String name;
     private final String surname;
     private final String specialization;
-    private final Long studioId;
+    private final Integer studioId;
 
-    public DevDTO(Long id, String name, String surname, String specialization, Long studioId) {
+    public DevDTO(Integer id, String name, String surname, String specialization, Integer studioId) {
         this.id = id;
         this.name = name;
         this.surname = surname;
@@ -15,7 +15,7 @@ public class DevDTO {
         this.studioId = studioId;
     }
 
-    public Long getId() { return id;}
+    public Integer getId() { return id;}
 
     public String getName() {
         return name;
@@ -29,7 +29,7 @@ public class DevDTO {
         return specialization;
     }
 
-    public Long getStudioId() {
+    public Integer getStudioId() {
         return studioId;
     }
 }
