+++ src/main/java/cz/cvut/fit/tjv/popovle1/semestral/crud_gamedev/repository/GameRepo.java
@@ -10,12 +10,12 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface GameRepo extends CrudRepository<Game, Long> {
-    Optional<Game> findById(Long id);
+public interface GameRepo extends CrudRepository<Game, Integer> {
+    Optional<Game> findById(Integer id);
 
     Optional<Game> findByName(String name);
 
     Collection<Game> findAll();
 
-    List<Game> findByIdIn(List<Long> gamesIds);
+    List<Game> findByIdIn(List<Integer> gamesIds);
 }
